+++ Cargo.toml
@@ -10,6 +10,13 @@ edition = "2021"
 keywords = ["egui", "ui", "graph", "node-graph"]
 categories = ["gui", "visualization"]
 
+[workspace]
+members = [
+  ".",
+  "demo-core",
+  "demo-web",
+]
+
 [dependencies]
 egui = { version = "0.32", default-features = false, features = [
   "persistence",
@@ -22,6 +29,7 @@ petgraph = { version = "0.8", default-features = false, features = [
   "serde-1",
 ] }
 serde = { version = "1.0", features = ["derive"] }
+instant = "0.1"
 
 crossbeam = { version = "0.8", optional = true }
 
@@ -56,7 +64,12 @@ bevy_egui = "0.36"
 serde_json = "1.0"
 ureq = { version = "2.9", default-features = true, features = ["json"] }
 criterion = { version = "0.5", features = ["html_reports"] }
+demo-core = { path = "demo-core" }
 
 [[bench]]
 name = "fruchterman_reingold"
 harness = false
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+# Enable JS RNG source for all wasm dependents via feature unification.
+getrandom = { version = "0.2", features = ["js"] }
