+++ demo-core/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "demo-core"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+egui = "0.32"
+eframe = "0.32"
+egui_graphs = { path = ".." }
+petgraph = { version = "0.8", default-features = false, features = [
+  "graphmap",
+  "stable_graph",
+  "matrix_graph",
+] }
+rand = "0.9"
+instant = "0.1"
+
+[features]
+# Pass-through for consumers if needed (not used here)
+events = ["egui_graphs/events"]
