+++ demo-web/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "demo-web"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+console_error_panic_hook = "0.1"
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
+web-sys = { version = "0.3", features = ["Window", "Document", "HtmlCanvasElement"] }
+egui = "0.32"
+eframe = { version = "0.32" }
+demo-core = { path = "../demo-core" }
+egui_graphs = { path = ".." }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+getrandom = { version = "0.3", features = ["wasm_js"] }
+
+[package.metadata.wasm-pack.profile.dev]
+wasm-opt = false
