+++ demo-web/index.html
@@ -1,0 +1,36 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>egui_graphs demo</title>
+  <script type="importmap">
+    {
+      "imports": {
+        "env": "./env.js"
+      }
+    }
+  </script>
+  <style>
+    html, body {
+      margin: 0;
+      padding: 0;
+      height: 100%;
+      width: 100%;
+      overflow: hidden;
+    }
+    /* Let the canvas fill the viewport; eframe will set its pixel resolution. */
+    canvas#the_canvas_id {
+      width: 100vw;
+      height: 100vh;
+      display: block;
+      outline: none;
+    }
+  </style>
+  <link data-trunk rel="rust" data-cargo-manifest="./Cargo.toml" />
+  <link data-trunk rel="copy-file" href="env.js" />
+  </head>
+  <body>
+    <canvas id="the_canvas_id"></canvas>
+  </body>
+</html>
